Fix FindFFmpeg.cmake
The `PC_` prefix used with `pkg_check_module` was removed in 5a69a462c
but according to the comment this didn't make the code compile yet.

Looking at another place where this file can be found online [1] the
prefix should be there.

When `pkg_check_module` is called and a component is found the
`..._INCLUDE_DIRS` and `..._LIBRARIES` variables will be set. The
following `find_path` and `find_library` calls won't do anything because
because of that:

> [...] the result is stored in the variable and the search will not be
repeated [...] [2]

The `..._INCLUDE_DIRS` can be set to an empty string by
`pkg_check_module` which results in an empty `FFMPEG_INCLUDE_DIRS` which
makes `find_package_handle_standard_args` fail because the a required
variable is empty.

Building with `tools/linux/build-in-docker` works because ffmpeg is in
`/opt/ffmpeg` which makes `pkg_check_module` actually set a value for
`..._INCLUDE_DIRS`.

[1] https://github.com/snikulov/cmake-modules/blob/master/FindFFmpeg.cmake
[2] https://cmake.org/cmake/help/v3.8/command/find_path.html